Create new esbonio-extensions package (#146)
I think it makes more sense to distribute the language server and the extensions separately as their own package. This PR
- Converts the original `esbonio` package to a namespace package
- Creates a new `esbonio-extensions` package to hold any Sphinx extensions
