+++ README.md
@@ -5,5 +5,6 @@ This is a monorepo containing a number of subprojects
 
 | Folder | Description | Badges |
 |-|-|-|
-| `code/` | A VSCode for editing Sphinx documentation projects | [![Visual Studio Marketplace Version](https://img.shields.io/visual-studio-marketplace/v/swyddfa.esbonio?style=flat-square)![Visual Studio Marketplace Installs](https://img.shields.io/visual-studio-marketplace/i/swyddfa.esbonio?style=flat-square)![Visual Studio Marketplace Downloads](https://img.shields.io/visual-studio-marketplace/d/swyddfa.esbonio?style=flat-square)](https://marketplace.visualstudio.com/items?itemName=swyddfa.esbonio)|
-| `lib/esbonio/` | A collection of Sphinx extensions as well as a Language Server for Sphinx projects. | [![PyPI](https://img.shields.io/pypi/v/esbonio?style=flat-square)![PyPI - Downloads](https://img.shields.io/pypi/dm/esbonio?style=flat-square)](https://pypi.org/project/esbonio/) |
+| `code/` | A VSCode extension for editing Sphinx documentation projects | [![Visual Studio Marketplace Version](https://img.shields.io/visual-studio-marketplace/v/swyddfa.esbonio?style=flat-square)![Visual Studio Marketplace Installs](https://img.shields.io/visual-studio-marketplace/i/swyddfa.esbonio?style=flat-square)![Visual Studio Marketplace Downloads](https://img.shields.io/visual-studio-marketplace/d/swyddfa.esbonio?style=flat-square)](https://marketplace.visualstudio.com/items?itemName=swyddfa.esbonio)|
+| `lib/esbonio/` | A Language Server for Sphinx projects. | [![PyPI](https://img.shields.io/pypi/v/esbonio?style=flat-square)![PyPI - Downloads](https://img.shields.io/pypi/dm/esbonio?style=flat-square)](https://pypi.org/project/esbonio/) |
+| `lib/esbonio-extensions` | A collection of Sphinx extensions | |
