+++ lib/esbonio-extensions/.bumpversion.cfg
@@ -1,12 +1,10 @@
 [bumpversion]
-current_version = 0.5.1
+current_version = 0.1.0
 commit = False
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(.dev(?P<dev>\d+))?
 serialize =
 	{major}.{minor}.{patch}.dev{dev}
 	{major}.{minor}.{patch}
-
-[bumpversion:file:esbonio/lsp/__init__.py]
 
 [bumpversion:file:setup.cfg]
