+++ lib/esbonio-extensions/README.md
@@ -1,0 +1,5 @@
+# Esbonio Extensions
+
+**This package is in early development**
+
+A collection of Sphinx extensions.
