+++ lib/esbonio-extensions/changes/github-template.html
@@ -1,0 +1,3 @@
+%(body_pre_docinfo)s
+%(docinfo)s
+%(body)s
