+++ lib/esbonio-extensions/pyproject.toml
@@ -1,0 +1,64 @@
+[build-system]
+requires = ["setuptools >= 35.0.2", "wheel >= 0.29.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.towncrier]
+filename = "CHANGES.rst"
+directory = "changes/"
+title_format = "v{version} - {project_date}"
+issue_format = "`#{issue} <https://github.com/swyddfa/esbonio/issues/{issue}>`_"
+underlines = ["-", "^", "\""]
+
+[[tool.towncrier.type]]
+     directory = "feature"
+     name = "Features"
+     showcontent = true
+
+[[tool.towncrier.type]]
+     directory = "fix"
+     name = "Fixes"
+     showcontent = true
+
+[[tool.towncrier.type]]
+     directory = "doc"
+     name = "Docs"
+     showcontent = true
+
+[[tool.towncrier.type]]
+     directory = "breaking"
+     name = "Breaking Changes"
+     showcontent = true
+
+[[tool.towncrier.type]]
+     directory = "deprecated"
+     name = "Deprecated"
+     showcontent = true
+
+[[tool.towncrier.type]]
+     directory = "misc"
+     name = "Misc"
+     showcontent = true
+
+
+[tool.tox]
+legacy_tox_ini = """
+[tox]
+isolated_build = True
+
+[testenv]
+deps =
+    mock
+    pytest
+    pytest-cov
+extras = lsp
+commands =
+    pytest --cov=esbonio
+
+[testenv:pkg]
+deps =
+    wheel
+usedevelop = True
+commands =
+    python setup.py clean --all
+    python setup.py sdist bdist_wheel
+"""
