+++ lib/esbonio-extensions/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = esbonio
-version = 0.5.1
-description = A Language Server for Sphinx projects.
+name = esbonio-extensions
+version = 0.1.0
+description = A collection of Sphinx extensions
 long_description = file:README.md
 long_description_content_type = text/markdown
 author = Alex Carney
@@ -24,16 +24,14 @@ platforms = any
 [options]
 packages = find_namespace:
 install_requires =
-    appdirs
     sphinx
-    pygls<0.10.0
+    nbformat<5.1
 
 [options.packages.find]
-exclude = tests*
+exclude = tests
 
 [options.extras_require]
 dev = black ; flake8 ; pytest ; pytest-cov ; mock
-lsp = appdirs ; pygls<0.10.0
 
 [flake8]
 max-line-length = 88
