--- lib/esbonio/esbonio/__init__.py
@@ -1,3 +1,0 @@
-from ._version import __version__
-
-__all__ = ["__version__"]
