--- lib/esbonio/esbonio/_version.py
@@ -1,1 +1,0 @@
-__version__ = "0.5.1"
